# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a B2B-API server.",
        "title": "B2B Example API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8070",
    "basePath": "/api/v1",
    "paths": {
...
        "/merchant/get-rates": {
            "get": {
                "description": "GetExchangeRates",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Merchant"
                ],
                "summary": "GetExchangeRates route",
                "parameters": [
                    {
                        "type": "string",
                        "format": "UUID",
                        "description": " ",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetExchangeRatesResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/merchant/transactions": {
            "post": {
                "description": "GetHistoryOfTransactions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Merchant"
                ],
                "summary": "GetHistoryOfTransactions route",
                "parameters": [
                    {
                        "type": "string",
                        "format": "UUID",
                        "description": " ",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "format": "json",
                        "description": "history filters",
                        "name": "filters",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/merchant.FilterHistoryTransaction"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.HistoryOfTransactionsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "merchant.FilterHistoryTransaction": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "fromDate": {
                    "type": "string"
                },
                "pagination": {
                    "type": "object",
                    "properties": {
                        "count": {
                            "type": "integer"
                        },
                        "currentPage": {
                            "type": "integer"
                        }
                    }
                },
                "toDate": {
                    "type": "string"
                }
            }
        },
        "merchant.GroupOfTransactions": {
            "type": "object",
            "properties": {
                "groupDate": {
                    "type": "string"
                },
                "transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/merchant.Transaction"
                    }
                }
            }
        },
        "merchant.Rates": {
            "type": "object",
            "properties": {
                "buyRate": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "sellRate": {
                    "type": "string"
                }
            }
        },
...
        "merchant.Transaction": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "DateProcess",
                    "type": "string"
                },
                "documentNumber": {
                    "type": "string"
                },
                "fromAccount": {
                    "type": "object",
                    "properties": {
                        "amount": {
                            "type": "object",
                            "properties": {
                                "currencySymbol": {
                                    "type": "string"
                                },
                                "value": {
                                    "description": "Amount",
                                    "type": "string"
                                }
                            }
                        },
                        "imageName": {
                            "type": "string"
                        },
                        "number": {
                            "description": "Acc??Number",
                            "type": "string"
                        },
                        "title": {
                            "description": "senderName",
                            "type": "string"
                        }
                    }
                },
                "id": {
                    "description": "Это DocId",
                    "type": "string"
                },
                "paymentPurpose": {
                    "description": "nazn",
                    "type": "string"
                },
                "proceededAt": {
                    "type": "string"
                },
                "status": {
                    "description": "Это DocState",
                    "type": "string"
                },
                "toAccount": {
                    "type": "object",
                    "properties": {
                        "number": {
                            "description": "Acc??Number",
                            "type": "string"
                        },
                        "title": {
                            "description": "RecipientName",
                            "type": "string"
                        }
                    }
                },
                "type": {
                    "description": "TransType",
                    "type": "string"
                }
            }
        },
...
        "models.ErrorModel": {
            "type": "object",
            "properties": {
                "errorCode": {
                    "type": "integer"
                },
                "errorDescription": {
                    "type": "string"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "errorCode": {
                    "type": "integer"
                },
                "errorDescription": {
                    "type": "string"
                },
                "response": {}
            }
        },
...
        "models.GetExchangeRatesResponse": {
            "type": "object",
            "properties": {
                "base": {
                    "type": "string"
                },
                "msg": {
                    "$ref": "#/definitions/models.ErrorModel"
                },
                "rates": {
                    "$ref": "#/definitions/merchant.Rates"
                }
            }
        },
...
        "models.HistoryOfTransactionsResponse": {
            "type": "object",
            "properties": {
                "groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/merchant.GroupOfTransactions"
                    }
                },
                "msg": {
                    "$ref": "#/definitions/models.ErrorModel"
                }
            }
        },
...
    }
}